+++ user-api/src/main/java/com/zerobase/GameSell/user/controller/UserController.java
@@ -5,7 +5,7 @@ import static com.zerobase.GameSell.user.exception.ErrorCode.NOT_FOUND_USER;
 import com.user.gamedomain.config.JwtAuthenticationProvider;
 import com.user.gamedomain.domain.Dto.UserDto;
 import com.user.gamedomain.domain.common.UserVo;
-import com.zerobase.GameSell.user.domain.UserDtoMapper;
+import com.zerobase.GameSell.user.domain.repository.UserDtoMapper;
 import com.zerobase.GameSell.user.domain.model.User;
 import com.zerobase.GameSell.user.exception.UserException;
 import com.zerobase.GameSell.user.service.UserService;
